fix(engine): cleanup objects ctors.

+++ inc/com/centreon/engine/host.hh
@@ -103,8 +103,8 @@ class                host : public notifier {
                            std::string const& icon_image_alt,
                            std::string const& vrml_image,
                            std::string const& statusmap_image,
-                           int x_2d,
-                           int y_2d,
+                           double x_2d,
+                           double y_2d,
                            bool have_2d_coords,
                            double x_3d,
                            double y_3d,
@@ -212,16 +212,16 @@ class                host : public notifier {
   void               set_have_2d_coords(bool has_coords);
   bool               get_have_3d_coords() const;
   void               set_have_3d_coords(bool has_coords);
-  int                get_x_2d() const;
-  void               set_x_2d(int x);
-  int                get_y_2d() const;
-  void               set_y_2d(int y);
-  int                get_x_3d() const;
-  void               set_x_3d(int x);
-  int                get_y_3d() const;
-  void               set_y_3d(int y);
-  int                get_z_3d() const;
-  void               set_z_3d(int z);
+  double             get_x_2d() const;
+  void               set_x_2d(double x);
+  double             get_y_2d() const;
+  void               set_y_2d(double y);
+  double             get_x_3d() const;
+  void               set_x_3d(double x);
+  double             get_y_3d() const;
+  void               set_y_3d(double y);
+  double             get_z_3d() const;
+  void               set_z_3d(double z);
   int                get_should_be_drawn() const;
   void               set_should_be_drawn(int should_be_drawn);
   int                get_acknowledgement_type() const;
@@ -299,11 +299,11 @@ private:
   std::string         _statusmap_image;
   bool                _have_2d_coords;
   bool                _have_3d_coords;
-  int                 _x_2d;
-  int                 _y_2d;
-  int                 _x_3d;
-  int                 _y_3d;
-  int                 _z_3d;
+  double              _x_2d;
+  double              _y_2d;
+  double              _x_3d;
+  double              _y_3d;
+  double              _z_3d;
   int                 _should_be_drawn;
   int                 _acknowledgement_type;
   bool                _should_reschedule_current_check;
